arxiv.sh: remove natwidth and natheight
Without them, the manuscript does not build locally:

    ! LaTeX Error: Cannot determine size of graphic in figure-1/figure-config1-labe
    l.png (no BoundingBox).

To fix it, I added natwidth and natheight as described here:

    https://tex.stackexchange.com/a/17737

But with them in place, arXiv's processor dies with:

    ! Undefined control sequence.
    <argument> ...@base \Gin@ext image\GPT@AttrShort
    \ifx \GPT@print \ltx@empty...
    l.977 ...t=713]{figure-1/figure-config1-label.png}

    ?
    ! Emergency stop.
    <argument> ...@base \Gin@ext image\GPT@AttrShort
    \ifx \GPT@print \ltx@empty...
    l.977 ...t=713]{figure-1/figure-config1-label.png}

So my solution is to strip them out when bundling the arXiv archive.

Thanks to Jim the arXiv admin for knowing how to fix this!